Corregido error en Controlador para guardar nuevos clientes por campo foto en null. Cambiado el tipo del campo precio de Double a Float en Producto.

+++ src/main/java/com/example/springboot/datajpa/app/models/entity/ItemFactura.java
@@ -20,7 +20,7 @@ public class ItemFactura implements Serializable {
   private Producto producto;
 
   public Double calcularImporte() {
-    return cantidad * producto.getPrecio();
+    return cantidad * producto.getPrecio().doubleValue();
   }
 
   public Long getId() {
